Move converters to own file
